Added SparseLongArray, SparseFloatArray, SparseDoubleArray, SparseByteArray, DoubleArrayLiist, CharArrayList, ByteArrayList, BooleanArrayList. Fixed LongArrayList, IntArrayList. Added SparseArrayUtils.valuesToList and a few more SparseArrayUtils.toString.

+++ src/org/andengine/util/adt/list/IIntList.java
@@ -16,10 +16,10 @@ public interface IIntList {
 	// ===========================================================
 
 	public boolean isEmpty();
-	public float get(final int pIndex) throws ArrayIndexOutOfBoundsException;
+	public int get(final int pIndex) throws ArrayIndexOutOfBoundsException;
 	public void add(final int pItem);
 	public void add(final int pIndex, final int pItem) throws ArrayIndexOutOfBoundsException;
-	public float remove(final int pIndex) throws ArrayIndexOutOfBoundsException;
+	public int remove(final int pIndex) throws ArrayIndexOutOfBoundsException;
 	public int size();
 	public void clear();
 	public int[] toArray();
